+++ hash/tests/sha3-256_tester.c
@@ -48,9 +48,9 @@ static int _sha3_256_tester(const struct lc_hash *sha3_256, const char *name)
 	LC_HASH_CTX_ON_STACK(ctx256, sha3_256);
 	LC_SHA3_256_CTX_ON_STACK(ctx256_stack);
 
-	printf("hash ctx %s (%s implementation) len %lu\n", name,
+	printf("hash ctx %s (%s implementation) len %u\n", name,
 	       sha3_256 == lc_sha3_256_c ? "C" : "accelerated",
-	       LC_HASH_CTX_SIZE(sha3_256));
+	       (unsigned int)LC_HASH_CTX_SIZE(sha3_256));
 	lc_hash_init(ctx256);
 	lc_hash_update(ctx256, msg_256, 3);
 	lc_hash_final(ctx256, act);
